requirements_windows.txt
absl-py==1.0.0
anyio==3.4.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==21.4.0
audioread==2.1.9
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bleach==4.1.0
blis==0.7.5
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
comtypes==1.1.10
cymem==2.0.6
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
docopt==0.6.2
emoji==1.6.1
entrypoints==0.3
Flask==2.0.2
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gTTS==2.2.3
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
Js2Py==0.71
json5==0.9.6
jsonschema==4.3.2
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
langcodes==3.3.0
libclang==12.0.0
librosa==0.8.1
llvmlite==0.37.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.6
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.7
notebook==6.4.6
numba==0.54.1
numpy==1.20.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
pickleshare==0.7.5
pipwin==0.5.1
pooch==1.5.2
preshed==3.0.6
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pydub==0.25.1
Pygments==2.10.0
pyjsparser==2.7.1
pyparsing==3.0.6
pypiwin32==223
PyPrind==2.11.3
pyrsistent==0.18.0
pySmartDL==1.3.4
python-dateutil==2.8.2
python-dotenv==0.19.2
pyttsx3==2.90
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==1.1.6
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
SoundFile==0.10.3.post1
soupsieve==2.3.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
spacy-stanza==1.0.1
speech==0.5.2
SpeechRecognition==3.8.1
srsly==2.4.2
stanza==1.3.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thinc==8.0.13
threadpoolctl==3.0.0
torch==1.10.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typer==0.4.0
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wasabi==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
zipp==3.6.0